+++ src/process/mod.rs
@@ -1,6 +1,7 @@
 //! Defines how rules can process and mutate Lua nodes.
 
 mod evaluator;
+mod expression_serializer;
 #[cfg(test)]
 mod node_counter;
 mod node_processor;
@@ -10,6 +11,7 @@ pub(crate) mod utils;
 mod visitors;
 
 pub use evaluator::*;
+pub(crate) use expression_serializer::*;
 #[cfg(test)]
 pub use node_counter::NodeCounter;
 pub use node_processor::NodeProcessor;